requirements.txt.txt
dspy-ai>=2.6.0
torch>=2.2.0
transformers>=4.43.0
accelerate>=0.33.0
sentence-transformers>=3.0.0
bert-score>=0.3.13
evaluate>=0.4.2
detoxify>=0.5.2
textstat>=0.7.4
scikit-learn>=1.3.0
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.26.0
tqdm>=4.66.0
